[viewer]分离组件查看代码
